Allow union types to have additional fields, support required fields. (#43)
- Union types can now have additional fields beyond just `type` and `value`. These can be passed to each specialized constructor along with the value.
- Adds support for required fields in schemas, which essentially just makes the parameters for these fields in the constructors be required. These are not intended to be widely used, but there are certain things which are truly required.
- Uses these features to add a required `id` field to request types, and `requestId` field to response types.
- These will be used to support multiple ongoing requests and remove hacks with global variables for cfe/analyzer internals in the future.
- Also dropped the special `_dartWrapperToJson` function - afaik this was never necessary (but please let me know if that is wrong).

If you would like, I could separate out the support here from adding the actual required fields.

## Risks

This makes requests non-hermetic, because of the `id` field. This is always only in the _outer_ request type though. We can still do caching based on the _inner_ request type, which I think should work out fine.

Note that having the IDs on the outer types means they are not generally visible to the user APIs which is good.