Use cached results based on hash of query responses (#132)
Implements caching into the MacroHost class. This is not necessarily the most efficient place to do it, but it is the most general place.

Adds methods on QueryService to start and stop tracking queries. This will need to get more sophisticated later on, to support concurrent macros.

Currently, likely due to the poor performance of the hashing, this is essentially equivalent to no smart invalidation at all, maybe even slower by around 10%. As such we should consider whether or not we want to actually land this in its current state. I would probably lean towards landing it as is though, and then trying to iterate on faster hashing solutions afterwords.